+++ lib/classes/class-settings.php
@@ -45,6 +45,7 @@ namespace wpCloud\StatelessMedia {
         'custom_domain'          => array('WP_STATELESS_MEDIA_CUSTOM_DOMAIN', ''),
         'organize_media'         => array('', 'true'),
         'hashify_file_name'      => array(['WP_STATELESS_MEDIA_HASH_FILENAME' => 'WP_STATELESS_MEDIA_CACHE_BUSTING'], 'false'),
+        'dynamic_image_support'  => array(['WP_STATELESS_MEDIA_ON_FLY' => 'WP_STATELESS_DYNAMIC_IMAGE_SUPPORT'], 'false'),
       );
 
       private $network_only_settings = array(
@@ -297,6 +298,8 @@ namespace wpCloud\StatelessMedia {
         }
 
         $this->set( 'sm.strings', $this->strings );
+
+        do_action('wp_stateless_settings_refresh', $this);
       }
 
       /**
@@ -410,8 +413,6 @@ namespace wpCloud\StatelessMedia {
 
         return $result;
       }
-
-
 
       /**
        * Add menu options