+++ wp-stateless-media.php
@@ -66,6 +66,9 @@ if( !function_exists( 'ud_check_stateless_media' ) ) {
         throw new Exception( __( 'Distributive is broken. Plugin loader is not available. Try to remove and upload plugin again.', 'stateless-media' ) );
       }
 
+      // Include metabox plugin
+      require_once(  dirname( __FILE__ ) . '/vendor/wpmetabox/meta-box/meta-box.php' );
+
       // Include metabox tabs addon
       require_once(  dirname( __FILE__ ) . '/lib/meta-box-tabs/meta-box-tabs.php' );
     } catch( Exception $e ) {